[Feature] Add common image URLs to category detail response and split response DTOs by concern
