+++ pida-core/core-api/src/main/kotlin/com/pida/presentation/v2/category/response/category/MapCategoryResponse.kt
@@ -1,22 +1,8 @@
-package com.pida.presentation.v2.category.response
+package com.pida.presentation.v2.category.response.category
 
 import com.pida.category.CategoryLabel
 import com.pida.category.MapCategory
-import io.swagger.v3.oas.annotations.media.ArraySchema
 import io.swagger.v3.oas.annotations.media.Schema
-
-@Schema(description = "카테고리 목록 응답")
-data class MapCategoryAllResponse(
-    @field:ArraySchema(
-        schema = Schema(implementation = MapCategoryResponse::class),
-        arraySchema = Schema(description = "카테고리 목록"),
-    )
-    val list: List<MapCategoryResponse>,
-) {
-    companion object {
-        fun of(list: List<MapCategoryResponse>) = MapCategoryAllResponse(list)
-    }
-}
 
 @Schema(description = "카테고리 응답")
 data class MapCategoryResponse(
