+++ pida-core/core-domain/src/main/kotlin/com/pida/category/item/model/MapCategoryItem.kt
@@ -2,6 +2,7 @@ package com.pida.category.item.model
 
 import com.pida.blooming.BloomingStatus
 import com.pida.category.badge.model.MapCategoryBadge
+import com.pida.flowerspot.FlowerSpotImage
 import com.pida.support.geo.GeoJson
 import com.pida.support.geo.Region
 import java.time.LocalDate
@@ -15,6 +16,7 @@ data class MapCategoryItem(
     val geom: GeoJson? = null,
     val pinPoint: GeoJson,
     val region: Region,
+    val imageUrls: List<FlowerSpotImage> = emptyList(),
     val homepageUrl: String? = null,
     val mapUrl: String? = null,
     val startDate: LocalDate? = null,
